+++ .copier-answers.yml
@@ -1,15 +1,19 @@
 # Changes here will be overwritten by Copier; NEVER EDIT MANUALLY
-_commit: v4.2.3
+_commit: v4.2.5
 _src_path: https://github.com/jupyterlab/extension-template
 author_email: aflewandowski@alaska.edu
 author_name: Alex Lewandowski
+data_format: string
+file_extension: ''
 has_binder: true
 has_settings: false
-kind: server
+kind: frontend
 labextension_name: jupyterlab-jupyterbook-navigation
-project_short_description: A JupyterLab extension that mimics jupyter-book chapter
-    navigation on an un-built, cloned jupyter book in JupyterLab.
+mimetype: ''
+mimetype_name: ''
+project_short_description: "A JupyterLab extension that mimics jupyter-book chapter navigation on an un-built, cloned jupyter book in JupyterLab."
 python_name: jupyterlab_jupyterbook_navigation
 repository: https://github.com/ASFOpenSARlab/jupyterlab-jupyterbook-navigation
 test: true
+viewer_name: ''
