+++ README.md
@@ -37,22 +37,6 @@ To remove the extension, execute:
 pip uninstall jupyterlab_jupyterbook_navigation
 ```
 
-## Troubleshoot
-
-If you are seeing the frontend extension, but it is not working, check
-that the server extension is enabled:
-
-```bash
-jupyter server extension list
-```
-
-If the server extension is installed and enabled, but you are not seeing
-the frontend extension, check the frontend extension is installed:
-
-```bash
-jupyter labextension list
-```
-
 ## Contributing
 
 ### Development install
@@ -67,11 +51,9 @@ The `jlpm` command is JupyterLab's pinned version of
 # Clone the repo to your local environment
 # Change directory to the jupyterlab_jupyterbook_navigation directory
 # Install package in development mode
-pip install -e ".[test]"
+pip install -e "."
 # Link your development version of the extension with JupyterLab
 jupyter labextension develop . --overwrite
-# Server extension must be manually installed in develop mode
-jupyter server extension enable jupyterlab_jupyterbook_navigation
 # Rebuild extension Typescript source after making changes
 jlpm build
 ```
@@ -96,8 +78,6 @@ jupyter lab build --minimize=False
 ### Development uninstall
 
 ```bash
-# Server extension must be manually disabled in develop mode
-jupyter server extension disable jupyterlab_jupyterbook_navigation
 pip uninstall jupyterlab_jupyterbook_navigation
 ```
 
@@ -106,24 +86,6 @@ command. To find its location, you can run `jupyter labextension list` to figure
 folder is located. Then you can remove the symlink named `jupyterlab-jupyterbook-navigation` within that folder.
 
 ### Testing the extension
-
-#### Server tests
-
-This extension is using [Pytest](https://docs.pytest.org/) for Python code testing.
-
-Install test dependencies (needed only once):
-
-```sh
-pip install -e ".[test]"
-# Each time you install the Python package, you need to restore the front-end extension link
-jupyter labextension develop . --overwrite
-```
-
-To execute them, run:
-
-```sh
-pytest -vv -r ap --cov jupyterlab_jupyterbook_navigation
-```
 
 #### Frontend tests
