+++ binder/postBuild
@@ -15,11 +15,10 @@ from pathlib import Path
 
 ROOT = Path.cwd()
 
-
 def _(*args, **kwargs):
-    """Run a command, echoing the args
+    """ Run a command, echoing the args
 
-    fails hard if something goes wrong
+        fails hard if something goes wrong
     """
     print("\n\t", " ".join(args), "\n")
     return_code = subprocess.call(args, **kwargs)
@@ -27,22 +26,12 @@ def _(*args, **kwargs):
         print("\nERROR", return_code, " ".join(args))
         sys.exit(return_code)
 
-
 # verify the environment is self-consistent before even starting
 _(sys.executable, "-m", "pip", "check")
 
 # install the labextension
 _(sys.executable, "-m", "pip", "install", "-e", ".")
 _(sys.executable, "-m", "jupyter", "labextension", "develop", "--overwrite", ".")
-_(
-    sys.executable,
-    "-m",
-    "jupyter",
-    "server",
-    "extension",
-    "enable",
-    "jupyterlab_jupyterbook_navigation",
-)
 
 # verify the environment the extension didn't break anything
 _(sys.executable, "-m", "pip", "check")
@@ -56,5 +45,3 @@ _("jupyter", "labextension", "list")
 
 print("JupyterLab with jupyterlab_jupyterbook_navigation is ready to run with:\n")
 print("\tjupyter lab\n")
-
-_("jupyter-book", "create", "JUPYTER-BOOK")
