+++ jest.config.js
@@ -1,16 +1,16 @@
-const jestJupyterLab = require("@jupyterlab/testutils/lib/jest-config");
+const jestJupyterLab = require('@jupyterlab/testutils/lib/jest-config');
 
 const esModules = [
-  "@codemirror",
-  "@jupyter/ydoc",
-  "@jupyterlab/",
-  "lib0",
-  "nanoid",
-  "vscode-ws-jsonrpc",
-  "y-protocols",
-  "y-websocket",
-  "yjs"
-].join("|");
+  '@codemirror',
+  '@jupyter/ydoc',
+  '@jupyterlab/',
+  'lib0',
+  'nanoid',
+  'vscode-ws-jsonrpc',
+  'y-protocols',
+  'y-websocket',
+  'yjs'
+].join('|');
 
 const baseConfig = jestJupyterLab(__dirname);
 
@@ -18,11 +18,11 @@ module.exports = {
   ...baseConfig,
   automock: false,
   collectCoverageFrom: [
-    "src/**/*.{ts,tsx}",
-    "!src/**/*.d.ts",
-    "!src/**/.ipynb_checkpoints/*"
+    'src/**/*.{ts,tsx}',
+    '!src/**/*.d.ts',
+    '!src/**/.ipynb_checkpoints/*'
   ],
-  coverageReporters: ["lcov", "text"],
-  testRegex: "src/.*/.*.spec.ts[x]?$",
+  coverageReporters: ['lcov', 'text'],
+  testRegex: 'src/.*/.*.spec.ts[x]?$',
   transformIgnorePatterns: [`/node_modules/(?!${esModules}).+`]
 };
