--- jupyter-config/server-config/jupyterlab_jupyterbook_navigation.json
@@ -1,7 +1,0 @@
-{
-  "ServerApp": {
-    "jpserver_extensions": {
-      "jupyterlab_jupyterbook_navigation": true
-    }
-  }
-}
