+++ jupyterlab_jupyterbook_navigation/__init__.py
@@ -7,7 +7,6 @@ except ImportError:
     import warnings
     warnings.warn("Importing 'jupyterlab_jupyterbook_navigation' outside a proper installation.")
     __version__ = "dev"
-from .handlers import setup_handlers
 
 
 def _jupyter_labextension_paths():
@@ -15,22 +14,3 @@ def _jupyter_labextension_paths():
         "src": "labextension",
         "dest": "jupyterlab-jupyterbook-navigation"
     }]
-
-
-def _jupyter_server_extension_points():
-    return [{
-        "module": "jupyterlab_jupyterbook_navigation"
-    }]
-
-
-def _load_jupyter_server_extension(server_app):
-    """Registers the API handler to receive HTTP requests from the frontend extension.
-
-    Parameters
-    ----------
-    server_app: jupyterlab.labapp.LabApp
-        JupyterLab application instance
-    """
-    setup_handlers(server_app.web_app)
-    name = "jupyterlab_jupyterbook_navigation"
-    server_app.log.info(f"Registered {name} server extension")
