--- jupyterlab_jupyterbook_navigation/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Python unit tests for jupyterlab_jupyterbook_navigation."""
