+++ pyproject.toml
@@ -23,19 +23,8 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
 ]
 dependencies = [
-    "jupyter_server>=2.0.1,<3",
-    "sphinx-external-toc~=0.3.1"
 ]
 dynamic = ["version", "description", "authors", "urls", "keywords"]
-
-[project.optional-dependencies]
-test = [
-    "coverage",
-    "pytest",
-    "pytest-asyncio",
-    "pytest-cov",
-    "pytest-jupyter[server]>=0.6.0"
-]
 
 [tool.hatch.version]
 source = "nodejs"
@@ -50,7 +39,6 @@ exclude = [".github", "binder"]
 [tool.hatch.build.targets.wheel.shared-data]
 "jupyterlab_jupyterbook_navigation/labextension" = "share/jupyter/labextensions/jupyterlab-jupyterbook-navigation"
 "install.json" = "share/jupyter/labextensions/jupyterlab-jupyterbook-navigation/install.json"
-"jupyter-config/server-config" = "etc/jupyter/jupyter_server_config.d"
 
 [tool.hatch.build.hooks.version]
 path = "jupyterlab_jupyterbook_navigation/_version.py"
