+++ src/__tests__/jupyterlab_jupyterbook_navigation.spec.ts
@@ -2,8 +2,8 @@
  * Example of [Jest](https://jestjs.io/docs/getting-started) unit tests
  */
 
-describe("jupyterlab-jupyterbook-navigation", () => {
-  it("should be tested", () => {
+describe('jupyterlab-jupyterbook-navigation', () => {
+  it('should be tested', () => {
     expect(1 + 1).toEqual(2);
   });
 });
