+++ style/index.js
@@ -1,2 +1,2 @@
-import "./base.css";
+import './base.css';
 import "./index.css";
