+++ ui-tests/jupyter_server_test_config.py
@@ -4,13 +4,9 @@
 opens the server to the world and provide access to JupyterLab
 JavaScript objects through the global window variable.
 """
-from pathlib import Path
-
 from jupyterlab.galata import configure_jupyter_server
 
 configure_jupyter_server(c)
 
 # Uncomment to set server log level to debug level
 # c.ServerApp.log_level = "DEBUG"
-
-c.ServerApp.root_dir = str(Path.cwd())
