+++ ui-tests/playwright.config.js
@@ -1,13 +1,13 @@
 /**
  * Configuration for Playwright using default from @jupyterlab/galata
  */
-const baseConfig = require("@jupyterlab/galata/lib/playwright-config");
+const baseConfig = require('@jupyterlab/galata/lib/playwright-config');
 
 module.exports = {
   ...baseConfig,
   webServer: {
-    command: "jlpm start",
-    url: "http://localhost:8888/lab",
+    command: 'jlpm start',
+    url: 'http://localhost:8888/lab',
     timeout: 120 * 1000,
     reuseExistingServer: !process.env.CI
   }
