+++ ui-tests/tests/jupyterlab_jupyterbook_navigation.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "@jupyterlab/galata";
+import { expect, test } from '@jupyterlab/galata';
 
 /**
  * Don't load JupyterLab webpage before running the tests.
@@ -6,20 +6,16 @@ import { expect, test } from "@jupyterlab/galata";
  */
 test.use({ autoGoto: false });
 
-test("should emit an activation console message", async ({ page }) => {
+test('should emit an activation console message', async ({ page }) => {
   const logs: string[] = [];
 
-  page.on("console", message => {
+  page.on('console', message => {
     logs.push(message.text());
   });
 
   await page.goto();
 
   expect(
-    logs.filter(
-      s =>
-        s ===
-        "JupyterLab extension jupyterlab-jupyterbook-navigation is activated!"
-    )
+    logs.filter(s => s === 'JupyterLab extension jupyterlab-jupyterbook-navigation is activated!')
   ).toHaveLength(1);
 });
